update tauri cmds
