+++ src/tauri-cmds/os_folders/get_os_folders.ts
@@ -1,0 +1,9 @@
+import { invoke } from "@tauri-apps/api/core";
+import { OsFolder, SortType } from "../../models";
+
+export async function get_os_folders(userId: string, sort?: SortType) {
+  let sortType: SortType = "updated";
+  if (sort) { sortType = sort };
+  const osFolders: OsFolder[] = await invoke("get_os_folders", { userId, sortType });
+  return osFolders;
+}
